Implement PostgreSQL broker

+++ scripts/dev-extensions/build-docker-image.sh
@@ -1,4 +1,7 @@
 #!/bin/sh
 
+set -o errexit
 docker build -t ${SIDECAR_EXTENSION_IMAGE_NAME}:${SIDECAR_EXTENSION_IMAGE_TAG} --rm -f Dockerfile       .
-docker build -t ${SIDECAR_SETUP_IMAGE_NAME}:${SIDECAR_SETUP_IMAGE_TAG}         --rm -f Dockerfile-setup .
+if test -f Dockerfile-setup ; then
+    docker build -t ${SIDECAR_SETUP_IMAGE_NAME}:${SIDECAR_SETUP_IMAGE_TAG}     --rm -f Dockerfile-setup .
+fi
